+++ pricing-api/src/modules/banco_preco/filters/filters_operations.py
