+++ pricing-api/src/modules/banco_preco/filters/filters_repository.py
@@ -1,5 +1,5 @@
-from src.modules.filters.filters_operations import ListFiltersQueryParams
-from src.modules.items.item import ItemModel
+from src.modules.banco_preco.filters.filters_operations import ListFiltersQueryParams
+from src.modules.banco_preco.items.item import ItemModel
 from src.db.database import db_session, es
 from sqlalchemy import and_, desc, asc, func
