+++ pricing-api/src/modules/banco_preco/filters/filters_service.py
@@ -1,5 +1,5 @@
-from src.modules.filters.filters_operations import ListFiltersQueryParams
-from src.modules.filters.filters_repository import FilterRepository as repository
+from src.modules.banco_preco.filters.filters_operations import ListFiltersQueryParams
+from src.modules.banco_preco.filters.filters_repository import FilterRepository as repository
 
 class FilterService:
     def list(params: ListFiltersQueryParams):
