+++ pricing-api/src/modules/banco_preco/groups/group_schema.py
