+++ pricing-api/src/modules/banco_preco/groups/groups_operations.py
