+++ pricing-api/src/modules/banco_preco/items/item_schema.py
