+++ pricing-api/src/modules/banco_preco/pricing/pricing_operations.py
