+++ pricing-api/src/modules/banco_preco/pricing/pricing_repository.py
@@ -3,9 +3,9 @@ import os
 from dotenv import load_dotenv
 
 from src.db.database import es
-from src.modules.pricing.pricing_operations import PricingQuery
-from src.modules.utils.utils import Pageable
-from src.modules.utils.utils import get_princing_query
+from src.modules.banco_preco.pricing.pricing_operations import PricingQuery
+from src.modules.banco_preco.utils.utils import Pageable
+from src.modules.banco_preco.utils.utils import get_princing_query
 
 load_dotenv()
 ES_INDEX_ITEM = os.environ.get('ES_INDEX_ITEM')
