--- pricing-api/src/modules/charts/charts_service.py
@@ -1,9 +1,0 @@
-from typing import List
-from src.modules.charts.charts_operations import ChartsQueryParams
-from src.modules.charts.charts_repository import ChartsRepository as repository
-
-class ChartsService:
-
-    def get_aggregate(params: ChartsQueryParams):
-        values = repository.get_aggregate(params)
-        return values
